requirements.txt
streamlit==1.39.0
ultralytics==8.3.21
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy==1.26.4
pandas==2.2.3
torch>=2.0.0
